refactor(data): replace dump / populate with ingest
